feat: add notification component with mock data and styles
