+++ src/app/[year]/notification/(components)/data.ts
@@ -1,0 +1,160 @@
+//user profileImage 목업용 이미지입니다! 나중에 삭제해주시면 됩니당
+import mock_profile_img from "../../../../../public/assets/icons/notification/mock_user_profile.jpg";
+
+export const mockUsers = [
+  {
+    id: "user123",
+    nickname: "서머",
+    profileImage: mock_profile_img,
+  },
+  {
+    id: "user456",
+    nickname: "현머",
+    profileImage: mock_profile_img,
+  },
+  {
+    id: "user112",
+    nickname: "미머",
+    profileImage: mock_profile_img,
+  },
+  {
+    id: "user223",
+    nickname: "멈멈머머머머머머머",
+    profileImage: mock_profile_img,
+  },
+];
+
+export const mockNotifications = [
+  //내 소식 알림
+  {
+    id: 1,
+    userId: "user123",
+    type: "INVITE_GROUP_MOA",
+    message: "님이 새로운 그룹 모아에 초대했습니다.",
+    payload: "",
+    read: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 2,
+    userId: "user456",
+    type: "FRIEND_REQUEST",
+    message: "님이 친구 요청을 보냈습니다.",
+    payload: "",
+    read: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 21,
+    userId: "user112",
+    type: "FRIEND_REQUEST",
+    message: "님이 친구 요청을 보냈습니다.",
+    payload: "",
+    read: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 221,
+    userId: "user223",
+    type: "FRIEND_REQUEST",
+    message: "님이 친구 요청을 보냈습니다.",
+    payload: "",
+    read: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  //모아 소식 알림림
+  {
+    id: 3,
+    userId: "admin",
+    type: "LETTER_RECEIVED",
+    message: "우와! 편지 한 통을 받으셨어요!",
+    payload: "",
+    read: false,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 4,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 5,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 6,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 7,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 8,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 9,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 10,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+  {
+    id: 11,
+    userId: "admin",
+    type: "FROM_MOA",
+    message: "회원 가입을 축하드려요! 모아모아와 함께 추억을 모아봐요",
+    payload: "",
+    read: true,
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+  },
+];
