+++ src/styles/globals.css
@@ -29,7 +29,7 @@
   /* -------------------------- */
   /*  텍스트 색상상               */
   /* -------------------------- */
-  --text-primary: #0000 00; /* 기본 글자 색상*/
+  --text-primary: #000000; /* 기본 글자 색상*/
   --text-primary-reverse: #fff; /*  기본 글자 색상 전환(버튼에 쓰게) */
   --text-secondary: #0f0f0f; /* 보조 텍스트(회색 같은 거)*/
   --text-accent: #0f0f0f; /* 경고 텍스트 등 */
@@ -40,7 +40,7 @@
   --ui-primary: #333; /* 주 색상(버튼, 링크, 로고 등 주목도가 필요한 요소) */
   --ui-secondary: #fff; /* Primary 보조 및 강조 색상(디자인적 요소) */
   --ui-neutral: #000000; /* 텍스트, 배경, 테두리 등에 사용 */
-  --ui-accent: #000000; /* 특정 영역을 강조하거나 특별한 상황에 사용(ex 경고 문구등) */
+  --ui-accent: #ff6464; /* 특정 영역을 강조하거나 특별한 상황에 사용(ex 경고 문구등) */
 
   /* -------------------------- */
   /*  고려해볼만한 다른 공통 css     */
